module2/app/requirements.txt
confluent-kafka
faust-streaming
faust-streaming-rocksdb
rocksdict